Un-daemonized shutdown_button()
Putting shutdown_button() in a daemon thread made the shutdown process exit before it could actually run the shutdown command because when the server mode subprocess is killed, it exits out of the main program. Un-daemonizing the tread allows it to run the shutdown process, even when the main thread exits.